Standardize architecture naming across Linux and Mac
Some Linux/BSD systems use aarch64, others use arm64 to refer to the
underlying ARM architecture. The previous version of this branch
referred to that as aarch64. Given that there is some debate as to
what that architecture is called, it seems nicer to use the naming
convention that matches that used by Mac systems.